解决每次逆向生成xml文件追加导致服务无法启动的问题

+++ src/main/java/com/example/demo/mapper/UserMapper.java
@@ -9,7 +9,7 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface UserMapper {
-    int countByExample(UserExample example);
+    long countByExample(UserExample example);
 
     int deleteByExample(UserExample example);
